requirements.txt
# Core dependencies
# Microsoft Agent Framework (workflows + agents)
# Note: Agent Framework is currently published as beta pre-releases on PyPI.
# Pin to a known-good beta to keep installs reproducible.
agent-framework==1.0.0b251211

# Data modeling / validation
pydantic>=2.7.0

# Local configuration (.env) for CLI runs
python-dotenv>=1.0.1

# CLI
typer>=0.12.5

# Optional: better console output (Typer uses it when present)
rich>=13.9.0

# Simple local UI (run viewer + quiz)
streamlit>=1.37.0

